+++ src/devices/heiman.ts
@@ -2587,8 +2587,16 @@ export const definitions: DefinitionWithExtend[] = [
             heimanExtend.heimanClusterSpecial(),
             m.deviceEndpoints({endpoints: {l1: 1, l2: 2}}),
             m.onOff({endpointNames: ["l1", "l2"]}),
-            m.deviceTemperature(),
             m.identify(),
+            m.numeric({
+                name: "device_intenal_temperature",
+                cluster: "genDeviceTempCfg",
+                attribute: "currentTemperature",
+                description: "Temperature of the device",
+                unit: "°C",
+                access: "STATE_GET",
+                entityCategory: "diagnostic",
+            }),
             m.enumLookup<"heimanClusterSpecial", HeimanPrivateCluster>({
                 name: "switch_type_l1",
                 endpointName: "l1",
@@ -3644,16 +3652,6 @@ export const definitions: DefinitionWithExtend[] = [
             }),
             heimanExtend.heimanClusterRadarDetectionRange(),
             m.numeric<"heimanClusterSpecial", HeimanPrivateCluster>({
-                name: "repeated_reporting_duration",
-                unit: "minute(s)",
-                valueMin: 0,
-                valueMax: 65535,
-                cluster: "heimanClusterSpecial",
-                attribute: "occupiedToOccupiedDuration",
-                description: "occupied repeated reporting duartion, 65535 indicates forever",
-                access: "ALL",
-            }),
-            m.numeric<"heimanClusterSpecial", HeimanPrivateCluster>({
                 name: "illuminance_threshold",
                 unit: "lx",
                 scale: 1,
@@ -3682,9 +3680,10 @@ export const definitions: DefinitionWithExtend[] = [
             }),
             heimanExtend.temperatureOffset(),
             heimanExtend.humidityOffset(),
-            m.enumLookup<"heimanClusterSpecial", HeimanPrivateCluster>({
+            m.binary<"heimanClusterSpecial", HeimanPrivateCluster>({
                 name: "learning_control",
-                lookup: {start: 65534, reset: 65535},
+                valueOn: ["start", 65534],
+                valueOff: ["reset", 65535],
                 cluster: "heimanClusterSpecial",
                 attribute: "radarLearningControl",
                 description: "Radar learning mode, please wake up the device first.",
@@ -3692,7 +3691,7 @@ export const definitions: DefinitionWithExtend[] = [
             }),
             m.enumLookup<"heimanClusterSpecial", HeimanPrivateCluster>({
                 name: "learning_state",
-                lookup: {not: 0, learning: 1, completed: 2, failed: 3},
+                lookup: {normal: 0, learning: 1, completed: 2, failed: 3},
                 cluster: "heimanClusterSpecial",
                 attribute: "radarLearningState",
                 description: "radar learning state",
@@ -3707,6 +3706,59 @@ export const definitions: DefinitionWithExtend[] = [
                 description: "Enable/disable the indicator on product",
                 access: "ALL",
             }),
+            heimanExtend.reportedPackages(),
+            heimanExtend.rejoinedCount(),
+            heimanExtend.rebootedCount(),
+        ],
+        ota: true,
+    },
+    {
+        zigbeeModel: ["HM-5HA-E", "HM-5HA-EF-3.0"],
+        model: "HM-5HA-E",
+        vendor: "Heiman",
+        description: "Heat detector",
+        fromZigbee: [fzLocal.heimanClusterSpecialfz],
+        toZigbee: [tz.warning],
+        configure: async (device, coordinatorEndpoint) => {
+            const endpoint = device.getEndpoint(1);
+            await reporting.bind(endpoint, coordinatorEndpoint, ["genPowerCfg", "heimanClusterSpecial"]);
+            await reporting.batteryPercentageRemaining(endpoint);
+            await endpoint.read("ssIasZone", ["zoneStatus", "zoneState", "iasCieAddr", "zoneId"]);
+            await endpoint.read("msTemperatureMeasurement", ["measuredValue"]);
+            await endpoint.read<"heimanClusterSpecial", HeimanPrivateCluster>(
+                "heimanClusterSpecial",
+                [
+                    "sensorFaultState",
+                    "deviceMuteControl",
+                    "deviceMuteState",
+                    "indicatorLightLevelControlOf1",
+                    "interconnectable",
+                    "temperatureOffset",
+                    "rebootedCount",
+                    "rejoinedCount",
+                    "reportedPackages",
+                ],
+                {
+                    manufacturerCode: Zcl.ManufacturerCode.HEIMAN_TECHNOLOGY_CO_LTD,
+                },
+            );
+        },
+        exposes: [],
+        extend: [
+            m.battery(),
+            m.identify(),
+            m.temperature(),
+            m.iasZoneAlarm({zoneType: "alarm", zoneAttributes: ["alarm_1", "battery_low", "test", "tamper"]}),
+            heimanExtend.heimanClusterSpecial(),
+            heimanExtend.heimanClusterSensorFaultState(),
+            heimanExtend.heimanClusterDeviceMuteState(),
+            heimanExtend.iasZoneInitiateTestMode(),
+            heimanExtend.heimanClusterSensorMutable(),
+            heimanExtend.heimanClusterIndicatorLight(),
+            heimanExtend.heimanClusterSensorInterconnectable(),
+            heimanExtend.sirenForAutomationOnly(),
+            heimanExtend.temperatureOffset(),
+            heimanExtend.linkAvailable(),
             heimanExtend.reportedPackages(),
             heimanExtend.rejoinedCount(),
             heimanExtend.rebootedCount(),